Create reusable UserInfoBlock component
